pyproject.toml
[tool.poetry]
name = "dlt-resea"
version = "0.1.0"
description = ""
authors = ["Kevin Wilson <khwilson@gmail.com>"]
packages = [
    { include = "dlt", from = "src" }
]

[tool.poetry.scripts]
"dlt-gen-random" = "dlt.gen_random:main"

[tool.poetry.dependencies]
python = ">=3.8,<3.10"
jupyter = "^1.0.0"
pandas = "^1.3.2"
numpy = "^1.21.2"
scipy = "^1.7.1"
matplotlib = "^3.4.3"
statsmodels = "^0.12.2"
tqdm = "^4.62.3"
seaborn = "^0.11.2"
names = "^0.3.0"
python-dotenv = "^0.19.1"
openpyxl = "^3.0.9"
pandera = "^0.7.2"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
pylint = "^2.6.0"
pytest = "^6.2.1"
isort = "^5.7.0"
ipdb = "^0.13.4"
mypy = "^0.790"
pytest-asyncio = "^0.14.0"
nbqa = "^1.1.1"

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88

[tool.pylint.basic]
good-names = "i,j,k,ex,Run,_,df,pc"

[tool.pylint.messages_control]
disable = "C0330, C0326, R0912, R0913, R0914, R0915"

[tool.pylint.format]
max-line-length = "88"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
